deployment/requirements.txt
ultralytics
torch
torchvision
opencv-python
pillow
gradio
